- Split SubtitleColor into FormattableColor and SubtitleColor. - Consolidated color parsing with StringUtil method.
